Added Game over Condition and option to restart the game

+++ HealthBar.cpp
@@ -34,7 +34,7 @@ HealthBar::~HealthBar()
 	}
 }
 
-const float& HealthBar::getHp() const
+const int& HealthBar::getHp() const
 {
 	return this->numberOfLives;
 }
